👚🔆 Updated with Glitch

+++ features/preventTalkingBump.js
@@ -1,8 +1,9 @@
-module.exports = client => {
-    const bumpChannel = "1051694989677166622";
-    client.on("messageCreate", (message) => {
-        if (message.channel.id === bumpChannel) {
-            message.delete();
-        }
-    });
-}
+module.exports = (client) => {
+  const bumpChannel = "1051694989677166622";
+  client.on("messageCreate", (message) => {
+    if (message.author.bot) return false;
+    if (message.channel.id === bumpChannel) {
+      message.delete();
+    }
+  });
+};
